+++ src/engine/data_storage/column.h
@@ -23,7 +23,7 @@ public:
     virtual void ReadMf(Reader&) = 0;
     virtual void PrintElemCsv(Writer&, size_t, bool) const = 0;
     virtual std::string GetElemToString(size_t index) const = 0;
-    virtual void Accept(ColumnVisitor& visitor, const std::vector<size_t>& group_indices, size_t max_group_index) const = 0;
+    virtual void Accept(ColumnVisitor& visitor, const std::vector<uint32_t>& group_indices, size_t max_group_index) const = 0;
     virtual bool Compare(const std::string&, size_t, CompareSign) const = 0;
     virtual void Filter(const std::string& value, CompareSign sign, std::vector<bool>& banned) const = 0;
     virtual std::shared_ptr<Column> CopyFiltered(const std::vector<bool>& banned) const = 0;
@@ -238,7 +238,7 @@ public:
         }
     }
 
-    void Accept(ColumnVisitor& visitor, const std::vector<size_t>& group_indices, size_t max_group_index) const override {
+    void Accept(ColumnVisitor& visitor, const std::vector<uint32_t>& group_indices, size_t max_group_index) const override {
         visitor.Visit(*this, group_indices, max_group_index);
     }
 
@@ -393,7 +393,7 @@ public:
     void PrintElemCsv(Writer&, size_t, bool) const override;
     std::string GetElemToString(size_t index) const override;
     std::string_view GetElemView(size_t index) const;
-    void Accept(ColumnVisitor& visitor, const std::vector<size_t>& group_indices, size_t max_group_index) const override;
+    void Accept(ColumnVisitor& visitor, const std::vector<uint32_t>& group_indices, size_t max_group_index) const override;
     bool Compare(const std::string&, size_t, CompareSign) const override;
     void Filter(const std::string& value, CompareSign sign, std::vector<bool>& banned) const override;
     std::shared_ptr<Column> CopyFiltered(const std::vector<bool>& banned) const override;
@@ -424,7 +424,7 @@ public:
     void ReadMf(Reader&) override;
     void PrintElemCsv(Writer&, size_t, bool) const override;
     std::string GetElemToString(size_t index) const override;
-    void Accept(ColumnVisitor& visitor, const std::vector<size_t>& group_indices, size_t max_group_index) const override;
+    void Accept(ColumnVisitor& visitor, const std::vector<uint32_t>& group_indices, size_t max_group_index) const override;
     bool Compare(const std::string&, size_t, CompareSign) const override;
     void Filter(const std::string& value, CompareSign sign, std::vector<bool>& banned) const override;
     std::shared_ptr<Column> CopyFiltered(const std::vector<bool>& banned) const override;
@@ -453,7 +453,7 @@ public:
     void ReadMf(Reader&) override;
     void PrintElemCsv(Writer&, size_t, bool) const override;
     std::string GetElemToString(size_t index) const override;
-    void Accept(ColumnVisitor& visitor, const std::vector<size_t>& group_indices, size_t max_group_index) const override;
+    void Accept(ColumnVisitor& visitor, const std::vector<uint32_t>& group_indices, size_t max_group_index) const override;
     bool Compare(const std::string&, size_t, CompareSign) const override;
     void Filter(const std::string& value, CompareSign sign, std::vector<bool>& banned) const override;
     std::shared_ptr<Column> CopyFiltered(const std::vector<bool>& banned) const override;
